Remove unused DuckDBCacheService
The data is already in the memory. Cache the data from DuckDB into DuckDB is not sense.
