+++ accio-tests/src/test/java/io/accio/testing/bigquery/TestBigQueryCache.java
@@ -41,7 +41,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThatNoException
 
 @Test(singleThreaded = true)
 public class TestBigQueryCache
-        extends AbstractCacheTestWithBigQuery
+        extends AbstractCacheTest
 {
     @Override
     protected Optional<String> getAccioMDLPath()
