+++ accio-tests/src/test/java/io/accio/testing/bigquery/TestBigQueryCacheService.java
@@ -20,7 +20,6 @@ import io.accio.cache.CacheService;
 import io.accio.cache.PathInfo;
 import io.accio.main.connector.CacheServiceManager;
 import io.accio.main.metadata.Metadata;
-import io.accio.testing.AbstractCacheTest;
 import io.accio.testing.TestingAccioServer;
 import org.testng.annotations.Test;
