--- accio-tests/src/test/resources/duckdb/cache_mdl.json
@@ -1,217 +1,0 @@
-{
-  "catalog": "memory",
-  "schema": "tpch",
-  "models": [
-    {
-      "name": "Orders",
-      "refSql": "select * from memory.tpch.orders",
-      "columns": [
-        {
-          "name": "orderkey",
-          "expression": "o_orderkey",
-          "type": "integer"
-        },
-        {
-          "name": "custkey",
-          "expression": "o_custkey",
-          "type": "integer"
-        },
-        {
-          "name": "orderstatus",
-          "expression": "o_orderstatus",
-          "type": "string"
-        },
-        {
-          "name": "totalprice",
-          "expression": "o_totalprice",
-          "type": "float"
-        },
-        {
-          "name": "customer",
-          "type": "Customer",
-          "relationship": "OrdersCustomer"
-        },
-        {
-          "name": "orderdate",
-          "expression": "o_orderdate",
-          "type": "date"
-        }
-      ],
-      "primaryKey": "orderkey",
-      "cached": "true"
-    },
-    {
-      "name": "Customer",
-      "refSql": "select * from memory.tpch.customer",
-      "columns": [
-        {
-          "name": "custkey",
-          "expression": "c_custkey",
-          "type": "integer"
-        },
-        {
-          "name": "name",
-          "expression": "c_name",
-          "type": "string"
-        },
-        {
-          "name": "orders",
-          "type": "Orders",
-          "relationship": "OrdersCustomer"
-        }
-      ],
-      "primaryKey": "custkey"
-    },
-    {
-      "name": "RemoveCustomer",
-      "refSql": "select * from memory.tpch.customer",
-      "columns": [
-        {
-          "name": "custkey",
-          "expression": "c_custkey",
-          "type": "integer"
-        },
-        {
-          "name": "name",
-          "expression": "c_name",
-          "type": "string"
-        },
-        {
-          "name": "orders",
-          "type": "Orders",
-          "relationship": "OrdersCustomer"
-        }
-      ],
-      "primaryKey": "custkey",
-      "cached": "true"
-    }
-  ],
-  "relationships": [
-    {
-      "name": "OrdersCustomer",
-      "models": [
-        "Orders",
-        "Customer"
-      ],
-      "joinType": "MANY_TO_ONE",
-      "condition": "Orders.custkey = Customer.custkey"
-    }
-  ],
-  "metrics": [
-    {
-      "cached": "true",
-      "name": "Revenue",
-      "baseObject": "Orders",
-      "dimension": [
-        {
-          "name": "custkey",
-          "type": "integer"
-        }
-      ],
-      "measure": [
-        {
-          "name": "revenue",
-          "type": "integer",
-          "expression": "sum(totalprice)"
-        }
-      ],
-      "timeGrain": [
-        {
-          "name": "orderdate",
-          "refColumn": "orderdate",
-          "dateParts": [
-            "YEAR",
-            "MONTH"
-          ]
-        }
-      ]
-    },
-    {
-      "cached": "true",
-      "name": "unqualified",
-      "baseObject": "Orders",
-      "dimension": [
-        {
-          "name": "custkey",
-          "type": "integer"
-        },
-        {
-          "name": "fake",
-          "type": "integer"
-        }
-      ],
-      "measure": [
-        {
-          "name": "revenue",
-          "type": "integer",
-          "expression": "sum(totalprice)"
-        }
-      ],
-      "timeGrain": [
-        {
-          "name": "orderdate",
-          "refColumn": "orderdate",
-          "dateParts": [
-            "YEAR",
-            "MONTH"
-          ]
-        }
-      ]
-    },
-    {
-      "name": "AvgRevenue",
-      "baseObject": "Orders",
-      "dimension": [
-        {
-          "name": "custkey",
-          "type": "integer"
-        }
-      ],
-      "measure": [
-        {
-          "name": "avg_revenue",
-          "type": "double",
-          "expression": "avg(totalprice)"
-        }
-      ],
-      "timeGrain": [
-        {
-          "name": "orderdate",
-          "refColumn": "orderdate",
-          "dateParts": [
-            "YEAR",
-            "MONTH"
-          ]
-        }
-      ]
-    },
-    {
-      "cached": "true",
-      "name": "ForDropTable",
-      "baseObject": "Orders",
-      "dimension": [
-        {
-          "name": "custkey",
-          "type": "integer"
-        }
-      ],
-      "measure": [
-        {
-          "name": "revenue",
-          "type": "integer",
-          "expression": "sum(totalprice)"
-        }
-      ],
-      "timeGrain": [
-        {
-          "name": "orderdate",
-          "refColumn": "orderdate",
-          "dateParts": [
-            "YEAR",
-            "MONTH"
-          ]
-        }
-      ]
-    }
-  ]
-}
